Makefile: fix update_doc target failing to checkout the correct branch
When we clone from .git, we get refs from the local machine as well but
for updating docs we want gh-pages from the actual origin, not the local one.

The way it was before, it would either fail (if you don't have a
gh-pages branch locally) or do the wrong thing, using a possibly older
version of gh-pages.